bunch of clang warning fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@23425 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/dted/dted_ptstream.c
@@ -563,7 +563,7 @@ void DTEDPtStreamSetMetadata( void *hStream, DTEDMetaDataCode eCode,
 {
     DTEDPtStream *psStream = (DTEDPtStream *) hStream;
 
-    if( eCode >= 0 && eCode < DTEDMD_MAX+1 )
+    if( (int)eCode >= 0 && eCode < DTEDMD_MAX+1 )
     {
         CPLFree( psStream->apszMetadata[eCode] );
         psStream->apszMetadata[eCode] = CPLStrdup( pszValue );
